requirements.txt
Pillow
click
pydub
pypandoc
docx
python-docx
markdown
markdown2
fpdf
pyth
pytest